+++ backend/Proppy.API/Resources/EmployeeResource.cs
@@ -4,7 +4,7 @@ namespace Proppy.API.Resources
 {
     public class EmployeeResource
     {
-        public int ID {get; set;}
+        public long ID {get; set;}
         public string Name {get; set;}
         public string Phone_No {get; set;}
         public string Email{get; set;}
